- change distance type in DB to integer - add update distance in DB - add distance validation - don't update empty values from table

+++ backend/src/main/java/com/aitseb/hamster/dao/Activity.java
@@ -44,6 +44,6 @@ public class Activity {
     private int effort;
     private int elevation;
     private float speed;
-    private float distance;
+    private int distance;
     private String notes;
 }
